Refactor Room connection handling to use Direction enum and update RoomTest for connection validation
