+++ src/main/java/org/example/RoomConnections.java
@@ -1,22 +1,6 @@
 package org.example;
 
 public class RoomConnections {
-    public enum Direction {
-        NORTH("norden"),
-        SOUTH("süden"),
-        EAST("osten"),
-        WEST("westen");
-
-        private final String label;
-
-        Direction(String label) {
-            this.label = label;
-        }
-
-        public String label() {
-            return label;
-        }
-    }
     private final Room kerker;
     private final Room waffenkammer;
     private final Room bibliothek;
